Organized object package into purpose-specific package
